Document debconf interaction with dpkg-reconfigure
It's not obvious at all that dpkg-reconfigure is needed to apply these changes.
Nor is it obvious why in some cases it completely ignores debconf values.

These docs let people understand and potentially dig into scripts to find workarounds.

It would be nice to find a way to invoke the frontend to just execute the postinst configure step.